changed time variable types to uint32_t

+++ include/searchinfo.h
@@ -4,9 +4,9 @@
 class SearchInfo
 {
 public:
-	uint64_t startTime;
-	uint64_t stopTime;
-	uint64_t timeLimit;
+	uint32_t startTime;
+	uint32_t stopTime;
+	uint32_t timeLimit;
 
 	uint32_t depth;
 	uint32_t depthLimit;
